requirements.txt
Flask>=2.3.2
supabase>=1.0.0
numpy>=1.26.2
python-dotenv>=1.0.0
earthengine-api>=0.1.319
weasyprint>=59.0
folium>=0.16.0
requests>=2.31.0
